silence failed autologon reg checks and mising ccm namespace check to keep the error collection clean

+++ Boxstarter.Bootstrapper/Cleanup-Boxstarter.ps1
@@ -17,10 +17,11 @@ function Cleanup-Boxstarter {
             $promptToExit=$true
         }
         $winLogonKey="HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon"
-        Remove-ItemProperty -Path $winLogonKey -Name "DefaultUserName" -ErrorAction SilentlyContinue
-        Remove-ItemProperty -Path $winLogonKey -Name "DefaultDomainName" -ErrorAction SilentlyContinue
-        Remove-ItemProperty -Path $winLogonKey -Name "DefaultPassword" -ErrorAction SilentlyContinue
-        Remove-ItemProperty -Path $winLogonKey -Name "AutoAdminLogon" -ErrorAction SilentlyContinue
+        $winlogonProps = Get-ItemProperty -Path $winLogonKey
+        if($winlogonProps.DefaultUserName){Remove-ItemProperty -Path $winLogonKey -Name "DefaultUserName" -ErrorAction SilentlyContinue}
+        if($winlogonProps.DefaultDomainName){Remove-ItemProperty -Path $winLogonKey -Name "DefaultDomainName" -ErrorAction SilentlyContinue}
+        if($winlogonProps.DefaultPassword){Remove-ItemProperty -Path $winLogonKey -Name "DefaultPassword" -ErrorAction SilentlyContinue}
+        if($winlogonProps.AutoAdminLogon){Remove-ItemProperty -Path $winLogonKey -Name "AutoAdminLogon" -ErrorAction SilentlyContinue}
         Write-Debug "Cleaned up logon registry and restart file"
         if($promptToExit -or $KeepWindowOpen){
             Read-Host 'Type ENTER to exit'
